backend/requirements.txt
streamlit
langchain
faiss-cpu
yt-dlp
assemblyai
openai
python-dotenv
audio-transcript
